Moving tx pre-mempool validation to TxHandler to avoid circular deps
